requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.14.0
huggingface-hub>=0.16.0
python-dotenv>=1.0.0
torch-pruning>=1.3.0
